+++ packages/url_launcher/url_launcher_platform_interface/lib/url_launcher_platform_interface.dart
@@ -57,6 +57,7 @@ abstract class UrlLauncherPlatform extends PlatformInterface {
     required bool enableJavaScript,
     required bool enableDomStorage,
     required bool universalLinksOnly,
+    required bool newTask,
     required Map<String, String> headers,
     String? webOnlyWindowName,
   }) {