+++ packages/url_launcher/url_launcher_web/lib/url_launcher_web.dart
@@ -79,6 +79,7 @@ class UrlLauncherPlugin extends UrlLauncherPlatform {
     bool enableJavaScript = false,
     bool enableDomStorage = false,
     bool universalLinksOnly = false,
+    bool newTask = false,
     Map<String, String> headers = const <String, String>{},
     String? webOnlyWindowName,
   }) {